requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.7
Django==4.2.6
djangorestframework==3.14.0
filelock==3.12.4
identify==2.5.30
kombu==5.3.2
Markdown==3.5
nodeenv==1.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8